Update Lexer tests
